Avance en el mòdulo administración.

+++ Proto1/Controllers/CrudController.cs
@@ -17,7 +17,7 @@ namespace Proto1.Controllers
         public ActionResult Index()
         {
 
-            if (System.Web.HttpContext.Current.Session["user_name"] == null || System.Web.HttpContext.Current.Session["db_path"] == null)
+            if (System.Web.HttpContext.Current.Session["is_logged"] == null)
             {
 
                 return RedirectToAction("Index", "Login");
@@ -26,43 +26,55 @@ namespace Proto1.Controllers
             else
             {
 
-                String user_name = System.Web.HttpContext.Current.Session["user_name"].ToString();
-                String db_path = System.Web.HttpContext.Current.Session["db_path"].ToString();
+                int user_type = Int32.Parse(System.Web.HttpContext.Current.Session["user_type"].ToString());
 
-                DataTable dt = new DataTable();
-                string strConString = @"Data Source=.\SQLEXPRESS;Initial Catalog=" + db_path + ";Integrated Security=True";
-
-                using (SqlConnection con = new SqlConnection(strConString))
+                if (user_type == 1 || user_type == 2)
                 {
 
-                    con.Open();
-                    SqlCommand cmd = new SqlCommand("Select * from Entity", con);
-                    SqlDataAdapter da = new SqlDataAdapter(cmd);
-                    da.Fill(dt);
+                    String db_path = System.Web.HttpContext.Current.Session["db_path"].ToString();
+
+                    DataTable dt = new DataTable();
+                    string strConString = @"Data Source=.\SQLEXPRESS;Initial Catalog=" + db_path + ";Integrated Security=True";
+
+                    using (SqlConnection con = new SqlConnection(strConString))
+                    {
+
+                        con.Open();
+                        SqlCommand cmd = new SqlCommand("Select * from Entity", con);
+                        SqlDataAdapter da = new SqlDataAdapter(cmd);
+                        da.Fill(dt);
+
+                    }
+
+                    var entityList = new List<Entity>();
+
+                    foreach (DataRow row in dt.Rows)
+                    {
+
+                        int id = Int32.Parse(row["id"].ToString());
+                        String name = row["name"].ToString();
+                        String content = row["content"].ToString();
+                        DateTime creation_date = Convert.ToDateTime(row["creation_date"].ToString());
+
+                        Entity entity = new Entity();
+                        entity.id = id;
+                        entity.name = name;
+                        entity.content = content;
+                        entity.creation_date = creation_date;
+
+                        entityList.Add(entity);
+
+                    }
+
+                    return View(entityList);
 
                 }
-
-                var entityList = new List<Entity>();
-
-                foreach (DataRow row in dt.Rows)
+                else
                 {
 
-                    int id = Int32.Parse(row["id"].ToString());
-                    String name = row["name"].ToString();
-                    String content = row["content"].ToString();
-                    DateTime creation_date = Convert.ToDateTime(row["creation_date"].ToString());
-
-                    Entity entity = new Entity();
-                    entity.id = id;
-                    entity.name = name;
-                    entity.content = content;
-                    entity.creation_date = creation_date;
-
-                    entityList.Add(entity);
+                    return Content("No tiene permiso para ingresar aquí.");
 
                 }
-
-                return View(entityList);
 
             }
 
@@ -72,7 +84,7 @@ namespace Proto1.Controllers
         public ActionResult Create()
         {
 
-            if (System.Web.HttpContext.Current.Session["user_name"] == null || System.Web.HttpContext.Current.Session["db_path"] == null)
+            if (System.Web.HttpContext.Current.Session["is_logged"] == null)
             {
 
                 return RedirectToAction("Index", "Login");
@@ -81,8 +93,20 @@ namespace Proto1.Controllers
             else
             {
 
-                Debug.WriteLine("Create GET");
-                return View();
+                int user_type = Int32.Parse(System.Web.HttpContext.Current.Session["user_type"].ToString());
+
+                if (user_type == 1 || user_type == 2)
+                {
+
+                    return View();
+
+                }
+                else
+                {
+
+                    return Content("No tiene permiso para ingresar aquí.");
+
+                }
 
             }
 
@@ -92,36 +116,7 @@ namespace Proto1.Controllers
         public ActionResult Create(Entity entity)
         {
 
-            String user_name = System.Web.HttpContext.Current.Session["user_name"].ToString();
-            String db_path = System.Web.HttpContext.Current.Session["db_path"].ToString();
-
-            Debug.WriteLine("Create GET");
-            string strConString = @"Data Source=.\SQLEXPRESS;Initial Catalog=" +db_path+ ";Integrated Security=True";
-
-            using (SqlConnection con = new SqlConnection(strConString))
-            {
-
-                con.Open();
-
-                String query = "insert into Entity(name,content,creation_date) values(@name, @content, @creation_date)";
-                SqlCommand cmd = new SqlCommand(query, con);
-
-                cmd.Parameters.Add("@name", entity.name);
-                cmd.Parameters.Add("@content", entity.content);
-                cmd.Parameters.Add("@creation_date", DateTime.Now);
-                cmd.ExecuteNonQuery();
-
-            }
-
-            return RedirectToAction("Index", "Crud");
-
-        }
-
-        [HttpGet]
-        public ActionResult Update(String id)
-        {
-
-            if (System.Web.HttpContext.Current.Session["user_name"] == null || System.Web.HttpContext.Current.Session["db_path"] == null)
+            if (System.Web.HttpContext.Current.Session["is_logged"] == null)
             {
 
                 return RedirectToAction("Index", "Login");
@@ -130,10 +125,73 @@ namespace Proto1.Controllers
             else
             {
 
-                Debug.WriteLine("Update (GET), ID: " + id);
-                Entity entity = GetEntity(id);
-                return View(entity);
+                int user_type = Int32.Parse(System.Web.HttpContext.Current.Session["user_type"].ToString());
 
+                if (user_type == 1 || user_type == 2)
+                {
+
+                    String user_name = System.Web.HttpContext.Current.Session["user_name"].ToString();
+                    String db_path = System.Web.HttpContext.Current.Session["db_path"].ToString();
+
+                    string strConString = @"Data Source=.\SQLEXPRESS;Initial Catalog=" + db_path + ";Integrated Security=True";
+
+                    using (SqlConnection con = new SqlConnection(strConString))
+                    {
+
+                        con.Open();
+
+                        String query = "insert into Entity(name,content,creation_date) values(@name, @content, @creation_date)";
+                        SqlCommand cmd = new SqlCommand(query, con);
+
+                        cmd.Parameters.AddWithValue("@name", entity.name);
+                        cmd.Parameters.AddWithValue("@content", entity.content);
+                        cmd.Parameters.AddWithValue("@creation_date", DateTime.Now);
+                        cmd.ExecuteNonQuery();
+
+                    }
+
+                    return RedirectToAction("Index", "Crud");
+
+                }
+                else
+                {
+
+                    return Content("No tiene permiso para ingresar aquí.");
+
+                }
+
+            }
+
+        }
+
+        [HttpGet]
+        public ActionResult Update(String id)
+        {
+
+            if (System.Web.HttpContext.Current.Session["is_logged"] == null)
+            {
+
+                return RedirectToAction("Index", "Login");
+
+            }
+            else
+            {
+
+                int user_type = Int32.Parse(System.Web.HttpContext.Current.Session["user_type"].ToString());
+
+                if (user_type == 1 || user_type == 2)
+                {
+
+                    Entity entity = GetEntity(id);
+                    return View(entity);
+
+                }
+                else
+                {
+
+                    return Content("No tiene permiso para ingresar aquí.");
+
+                }
 
             }
 
@@ -143,38 +201,7 @@ namespace Proto1.Controllers
         public ActionResult Update(Entity entity)
         {
 
-            String user_name = System.Web.HttpContext.Current.Session["user_name"].ToString();
-            String db_path = System.Web.HttpContext.Current.Session["db_path"].ToString();
-
-            Debug.WriteLine("Update (POST)");
-
-            string strConString = @"Data Source=.\SQLEXPRESS;Initial Catalog=" + db_path + ";Integrated Security=True";
-
-            using (SqlConnection con = new SqlConnection(strConString))
-            {
-
-                con.Open();
-
-                String query = "update Entity set name = @name, content = @content, creation_date = @creation_date where id = @id";
-                SqlCommand cmd = new SqlCommand(query, con);
-
-                cmd.Parameters.Add("@id", entity.id);
-                cmd.Parameters.Add("@name", entity.name);
-                cmd.Parameters.Add("@content", entity.content);
-                cmd.Parameters.Add("@creation_date", DateTime.Now);
-                cmd.ExecuteNonQuery();
-
-            }
-
-            return RedirectToAction("Index", "Crud");
-
-        }
-
-        [HttpGet]
-        public ActionResult Delete(String id)
-        {
-
-            if (System.Web.HttpContext.Current.Session["user_name"] == null || System.Web.HttpContext.Current.Session["db_path"] == null)
+            if (System.Web.HttpContext.Current.Session["is_logged"] == null)
             {
 
                 return RedirectToAction("Index", "Login");
@@ -183,9 +210,76 @@ namespace Proto1.Controllers
             else
             {
 
-                Debug.WriteLine("Delete(GET), ID: " + id);
-                Entity entity = GetEntity(id);
-                return View(entity);
+                int user_type = Int32.Parse(System.Web.HttpContext.Current.Session["user_type"].ToString());
+
+                if (user_type == 1 || user_type == 2)
+                {
+
+                    String user_name = System.Web.HttpContext.Current.Session["user_name"].ToString();
+                    String db_path = System.Web.HttpContext.Current.Session["db_path"].ToString();
+
+                    Debug.WriteLine("Update (POST)");
+
+                    string strConString = @"Data Source=.\SQLEXPRESS;Initial Catalog=" + db_path + ";Integrated Security=True";
+
+                    using (SqlConnection con = new SqlConnection(strConString))
+                    {
+
+                        con.Open();
+
+                        String query = "update Entity set name = @name, content = @content, creation_date = @creation_date where id = @id";
+                        SqlCommand cmd = new SqlCommand(query, con);
+
+                        cmd.Parameters.AddWithValue("@id", entity.id);
+                        cmd.Parameters.AddWithValue("@name", entity.name);
+                        cmd.Parameters.AddWithValue("@content", entity.content);
+                        cmd.Parameters.AddWithValue("@creation_date", DateTime.Now);
+                        cmd.ExecuteNonQuery();
+
+                    }
+
+                    return RedirectToAction("Index", "Crud");
+
+                }
+                else
+                {
+
+                    return Content("No tiene permiso para ingresar aquí.");
+
+                }
+
+            }
+
+        }
+
+        [HttpGet]
+        public ActionResult Delete(String id)
+        {
+
+            if (System.Web.HttpContext.Current.Session["is_logged"] == null)
+            {
+
+                return RedirectToAction("Index", "Login");
+
+            }
+            else
+            {
+
+                int user_type = Int32.Parse(System.Web.HttpContext.Current.Session["user_type"].ToString());
+
+                if (user_type == 1 || user_type == 2)
+                {
+
+                    Entity entity = GetEntity(id);
+                    return View(entity);
+
+                }
+                else
+                {
+
+                    return Content("No tiene permiso para ingresar aquí.");
+
+                }
 
             }
 
@@ -195,27 +289,49 @@ namespace Proto1.Controllers
         public ActionResult Delete(Entity entity)
         {
 
-            String user_name = System.Web.HttpContext.Current.Session["user_name"].ToString();
-            String db_path = System.Web.HttpContext.Current.Session["db_path"].ToString();
-
-            Debug.WriteLine("Delete(POST), ID: " + entity.id);
-
-            string strConString = @"Data Source=.\SQLEXPRESS;Initial Catalog=" + db_path + ";Integrated Security=True";
-
-            using (SqlConnection con = new SqlConnection(strConString))
+            if (System.Web.HttpContext.Current.Session["is_logged"] == null)
             {
 
-                con.Open();
-
-                String query = "delete from Entity where id = @id";
-                SqlCommand cmd = new SqlCommand(query, con);
-
-                cmd.Parameters.Add("@id", entity.id);
-                cmd.ExecuteNonQuery();
+                return RedirectToAction("Index", "Login");
 
             }
+            else
+            {
 
-            return RedirectToAction("Index", "Crud");
+                int user_type = Int32.Parse(System.Web.HttpContext.Current.Session["user_type"].ToString());
+
+                if (user_type == 1 || user_type == 2)
+                {
+
+                    String user_name = System.Web.HttpContext.Current.Session["user_name"].ToString();
+                    String db_path = System.Web.HttpContext.Current.Session["db_path"].ToString();
+
+                    string strConString = @"Data Source=.\SQLEXPRESS;Initial Catalog=" + db_path + ";Integrated Security=True";
+
+                    using (SqlConnection con = new SqlConnection(strConString))
+                    {
+
+                        con.Open();
+
+                        String query = "delete from Entity where id = @id";
+                        SqlCommand cmd = new SqlCommand(query, con);
+
+                        cmd.Parameters.AddWithValue("@id", entity.id);
+                        cmd.ExecuteNonQuery();
+
+                    }
+
+                    return RedirectToAction("Index", "Crud");
+
+                }
+                else
+                {
+
+                    return Content("No tiene permiso para ingresar aquí.");
+
+                }
+
+            }
 
         }
 
@@ -223,10 +339,32 @@ namespace Proto1.Controllers
         public ActionResult Details(String id)
         {
 
-            Debug.WriteLine("Details ID: " + id);
-            Entity entity = GetEntity(id);
+            if (System.Web.HttpContext.Current.Session["is_logged"] == null)
+            {
 
-            return View(entity);
+                return RedirectToAction("Index", "Login");
+
+            }
+            else
+            {
+
+                int user_type = Int32.Parse(System.Web.HttpContext.Current.Session["user_type"].ToString());
+
+                if (user_type == 1 || user_type == 2)
+                {
+
+                    Entity entity = GetEntity(id);
+                    return View(entity);
+
+                }
+                else
+                {
+
+                    return Content("No tiene permiso para ingresar aquí.");
+
+                }
+
+            }
 
         }
 
@@ -236,12 +374,8 @@ namespace Proto1.Controllers
             String user_name = System.Web.HttpContext.Current.Session["user_name"].ToString();
             String db_path = System.Web.HttpContext.Current.Session["db_path"].ToString();
 
-            Debug.WriteLine("Get Entity");
-            Debug.WriteLine("ID: " + id);
-
             DataTable dt = new DataTable();
             string strConString = @"Data Source=.\SQLEXPRESS;Initial Catalog=" +db_path+";Integrated Security=True";
-            Debug.WriteLine(strConString);
 
             using (SqlConnection con = new SqlConnection(strConString))
             {
@@ -253,16 +387,9 @@ namespace Proto1.Controllers
 
             }
 
-            Debug.WriteLine(id);
-
             String name = dt.Rows[0]["name"].ToString();
-            Debug.WriteLine(name);
-
             String content = dt.Rows[0]["content"].ToString();
-            Debug.WriteLine(content);
-
             DateTime creation_date = Convert.ToDateTime(dt.Rows[0]["creation_date"].ToString());
-            Debug.WriteLine(creation_date);
 
             Entity entity = new Entity();
             entity.id = Int32.Parse(id);
